refactor(server): align scaffold with opentunnel
