+++ packages/server/src/definition/index.ts
@@ -1,0 +1,1 @@
+export { api } from "./api.js"
