+++ packages/server/src/index.ts
@@ -1,3 +1,3 @@
 export { openapi } from "./openapi.js"
-export { api } from "./spec/api.js"
+export { api } from "./definition/api.js"
 export type { OpenApiSpec, ServerApi } from "./types.js"
